Make user renaming trigger an update in the room data.
We could use the existing User::nameChanged signal - however, there's a problem with such connection. When we have many rooms with many users we'll have a combinatorial number of connections. To remedy that, we introduce a new signal on the Connection object level that will be triggered every time when any user is renamed. This doesn't help to reduce the number of fire-ups but at least reduces the number of connections to set up and keep. At the same time, the renamed user is passed inside the signal anyway so we don't lose anything by consolidating the signals.