Introduce ServerCall infrastructure
This commit adds the following classes (with extensive comments throughout the code):
RequestParams - encapsulates parameters needed to construct a network request (replacing apiPath(), query() and data() virtual stacks in BaseJob and its derivatives);
CallStatus - a simple structure that stores the server call status;
ServerCallSetupBase and ServerCallSetup<> - a template to derive from in order to produce new types of server calls; supplied with a couple of result adapter types that can be passed to ServerCallSetup<> as a type parameter;
ServerCallBase and ServerCall<> - the classes that encapsulate the server call itself: ServerCallBase does heavy lifting of sending network requests and dealing with signals, while ServerCall<> is a template wrapper around ServerCallBase that allows to work with the response from the server in a type-safe way.
This commit also contains a proof-of-concept conversion of PasswordLogin job to ServerCallSetup<> and rewires Connection to use the new implementation.