Provide a default avatar
If the user doesn't have a custom avatar set, show a default avatar
instead of nothing. This fixes a misalignment in the tableview
between users with and without a custom avatar.

Since libqmatrixclient doesn't recognize yet if a user is online or offline,
just use the `user-available` standard [1] icon for now.

[1]: https://standards.freedesktop.org/icon-naming-spec/icon-naming-spec-latest.htmli