ignore possible appendixes from content type
Currently libqmatrixclient fails checking the `Content-Type` header
when there is an appendix like "charset". That is allowed e.g. in
[rfc7231](https://tools.ietf.org/html/rfc7231#section-3.1.1.5))

One example is a Content-Type `application/json` vs `application/json;charset=UTF-8`
Setting of the charset appendis is currently not supported. It fails with
libqmatrixclient.jobs: "LoginJob" status 106 : "Incorrect content type of the response"

This PR aims to just drop that appendix as it is currently not handled somewhere else.