Fix qmake build failure at AppVeyor
Turned out libQuotient code base was not good for NMake inference rules
but qmake has been intelligent enough to detect that and disable before.
After the recent optimisations qmake's intelligence was not enough so
CONFIG += no_batch was not automatically added, leading to the linking
failure because NMake skipped compilation of all source files.