(DO NOT MERGE) Profile User::*ForRoom() methods
This adds benchmarking code to the paths involving the flimsy and
complicated machinery around getting and setting per-room names and
avatars (in fact, the same code is used for non-room profile data too),
as opposed to straightforward Room::getCurrentState() calls. The
benchmarking code slows down execution quite a bit (after a few minutes
of regular usage the number of nameForRoom() invocations may easily go
over 100'000, and every time the benchmark dumps the stats to the logs
after measurements) so DO NOT USE IN PRODUCTION.

As for the outcomes of benchmarking: as much as Room::getCurrentState()
itself proved to be very unoptimised (it ensures that a member event
exists per each room+user combination, making 1000's of stub member
events just after visiting Matrix HQ) the accumulated time in
nameForRoom() is still smaller than the current implementation due to
enourmously heavy setNameForRoom(), even taking into account that
setNameForRoom() is called less frequently. YMMV but even without
optimising Room::getCurrentState() the new implementation can be about
2-3 times faster; and if allocations get removed from getCurrentState()
it can be up to 10 times faster. The conclusion is clear: the new code
is much shorter, considerably faster and saves memory.