Room: fix breakage in internal member map
The change in 39830496 led to prev_content becoming a fallback not only
for displaying user names but also for storing them in the internal
member map, which is really not what was intended.

A lot of debug logging has been added - this will be moved to a new
logging category before merging.