EventItemBase: Allow adding custom data
A new field of std::any type is added that allows clients to "annotate"
any event item with arbitrary kind of data. This is mainly intended so
that clients could calculate certain information about the item (e.g.
special formatting depending on the event contents, or position) without
having to calculate this information every time it is visualised.
In case of Quaternion, the idea is to calculate the "spamminess" of
the event basing on the past activity of a given user in this room -
calculating it upon displaying each event is extremely heavyweight.