CMakeLists: fixed potential linking errors around quotient_common.h
quotient_common.h has Q_NAMESPACE but no own compilation unit, and moc
was not called on it either - using metaobject data on an enumeration
defined in that file leads to a linking error due to sharedMetaObject
not being defined. The fix makes so that the file is detected by automoc
with the respective definition being generated.

Cherry-picked from a83ec900 (0.6.x branch).