Room: isMember(); memberState() only needs user id
Room::memberJoinState() was only used to check if the user has joined
the room (it couldn't be used for anything else), meaning that its best
replacement is actually not memberState() but isMember() introduced
hereby. It's also better to pass user ids instead of User objects to
memberState() and isMember() since that is enough to check membership.

# Conflicts:
#	lib/room.cpp
#	lib/room.h