Bind read receipts to userIds, not to User* values
This reduces the surface interacting with the User class that eventually
will be split into LocalUser (most part) and RoomMember (a tiny wrapper
around the member data in a given room, used almost everywhere in Room
where User currently is).

Also: dropped a log message when the new receipt is at or behind
the old one as it causes a lot of noise in the logs.